move AccountResolver to generic place
